fix: moralis webhook onboarding
When setting up a new webhook, Moralis sends a test message. For some reason that test message is failing the authenticity check. This prevents the webhook from registering successfully. We are currently implementing a workaround that returns 200 immediately if txs and erc20Transfers are empty.

This commit also includes scripts for registering a new webhook with Moralis.